requirements.txt
bert-for-tf2==0.11.2
flask==1.1.1
flask-Bootstrap==3.3.7.1
flask-moment==0.9.0
flask-wtf==0.14.2
tensorflow==2.1.0

